Refactor paragraph utils
